Создает компонент Cities
